docs: update WebRTC jsdocs
Updates config comments and moves constants into the constants file
