+++ packages/transport-webrtc/src/muxer.ts
@@ -1,3 +1,4 @@
+import { MUXER_PROTOCOL } from './constants.js'
 import { createStream } from './stream.js'
 import { drainAndClose, nopSink, nopSource } from './util.js'
 import type { DataChannelOptions } from './index.js'
@@ -6,8 +7,6 @@ import type { ComponentLogger, Logger, Stream, CounterGroup, StreamMuxer, Stream
 import type { AbortOptions } from '@multiformats/multiaddr'
 import type { Source, Sink } from 'it-stream-types'
 import type { Uint8ArrayList } from 'uint8arraylist'
-
-const PROTOCOL = '/webrtc'
 
 export interface DataChannelMuxerFactoryInit {
   /**
@@ -55,7 +54,7 @@ export class DataChannelMuxerFactory implements StreamMuxerFactory {
     this.components = components
     this.peerConnection = init.peerConnection
     this.metrics = init.metrics
-    this.protocol = init.protocol ?? PROTOCOL
+    this.protocol = init.protocol ?? MUXER_PROTOCOL
     this.dataChannelOptions = init.dataChannelOptions ?? {}
     this.log = components.logger.forComponent('libp2p:webrtc:muxerfactory')
 
@@ -135,7 +134,7 @@ export class DataChannelMuxer implements StreamMuxer {
     this.logger = components.logger
     this.streams = init.streams.map(s => s.stream)
     this.peerConnection = init.peerConnection
-    this.protocol = init.protocol ?? PROTOCOL
+    this.protocol = init.protocol ?? MUXER_PROTOCOL
     this.metrics = init.metrics
     this.dataChannelOptions = init.dataChannelOptions ?? {}
